+++ composer.json
@@ -23,5 +23,6 @@
             "App\\" : "app/"
         }
     },
+    "prefer-stable": true,
     "minimum-stability": "dev"
 }
